requirements.txt
gspread==5.12.0
google-auth==2.29.0
python-dotenv==1.0.1
twilio
python-dotenv